+++ Include/STM32LibraryCollection/Uart.hpp
@@ -149,15 +149,20 @@ concept IsUartMessage =
  */
 template <IsWorkingMode WorkingModeT, __Internal::__IsUniqueTag UniqueTagT>
 class Uart {
-    using UartCallbackManagerT = __Internal::__CallbackManager<
-        UART_HandleTypeDef, UniqueTagT
+    using TransmitCompleteCallbackT = __Internal::__CallbackManager<
+        UART_HandleTypeDef, UniqueTagT, STM32_UNIQUE_TAG,
+        HAL_UART_RegisterCallback, HAL_UART_UnRegisterCallback, HAL_UART_TX_COMPLETE_CB_ID
+    >;
+    using ReceiveCompleteCallbackT = __Internal::__CallbackManager<
+        UART_HandleTypeDef, UniqueTagT, STM32_UNIQUE_TAG,
+        HAL_UART_RegisterCallback, HAL_UART_UnRegisterCallback, HAL_UART_RX_COMPLETE_CB_ID
     >;
 public:
 
     /**
      * @typedef CallbackT, Type alias for callback functions.
      */
-    using CallbackT = UartCallbackManagerT::CallbackT;
+    using CallbackT = TransmitCompleteCallbackT::CallbackT;
 
     /**
      * @brief Construct Uart class.
@@ -165,19 +170,10 @@ public:
      * @param handle        Reference to the UART handle.
      */
     explicit Uart(UART_HandleTypeDef& handle) noexcept
-      : m_handle{handle}
-    {
-        HAL_UART_RegisterCallback(
-            &m_handle,
-            HAL_UART_TX_COMPLETE_CB_ID,
-            UartCallbackManagerT::TransmitCompleteCallback
-        );
-        HAL_UART_RegisterCallback(
-            &m_handle,
-            HAL_UART_RX_COMPLETE_CB_ID,
-            UartCallbackManagerT::ReceiveCompleteCallback
-        );
-    }
+      : m_handle{handle},
+        m_transmit_complete_callback{handle},
+        m_receive_complete_callback{handle}
+    { }
 
     /**
      * @defgroup Deleted copy and move members.
@@ -191,19 +187,10 @@ public:
 
     /**
      * @brief Destroy Uart class.
+     * 
+     * @note Callbacks are automatically unregistered via RAII.
      */
-    ~Uart()
-    {
-        HAL_UART_UnRegisterCallback(
-            &m_handle,
-            HAL_UART_TX_COMPLETE_CB_ID
-        );
-        HAL_UART_UnRegisterCallback(
-            &m_handle,
-            HAL_UART_RX_COMPLETE_CB_ID
-        );
-        UartCallbackManagerT::UnRegisterCallbacks();
-    }
+    ~Uart() = default;
 
     /**
      * @returns UART handle reference.
@@ -266,7 +253,7 @@ public:
     ) noexcept
     requires (!std::same_as<RxWorkingModeT, WorkingMode::Blocking>)
     {
-        UartCallbackManagerT::RegisterReceiveCompleteCallback(
+        m_receive_complete_callback.Set(
             std::move(complete_callback)
         );
         if constexpr (std::same_as<RxWorkingModeT, WorkingMode::Interrupt>){
@@ -338,7 +325,7 @@ public:
     ) noexcept
     requires (!std::same_as<TxWorkingModeT, WorkingMode::Blocking>)
     {
-        UartCallbackManagerT::RegisterTransmitCompleteCallback(
+        m_transmit_complete_callback.Set(
             std::move(complete_callback)
         );
         if constexpr (std::same_as<TxWorkingModeT, WorkingMode::Interrupt>){
@@ -362,6 +349,8 @@ public:
 
 private:
     UART_HandleTypeDef& m_handle;
+    TransmitCompleteCallbackT m_transmit_complete_callback;
+    ReceiveCompleteCallbackT m_receive_complete_callback;
 
     /**
      * @brief Clamp size to std::uint16_t range.